# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File Copy Manager API",
    "description": "API para gerenciamento de copia de arquivos com fila de processamento, dashboard web em tempo real e persistencia contra quedas de energia.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API local"
    }
  ],
...
  "paths": {
...
    "/buckets": {
      "get": {
        "tags": ["Buckets"],
        "summary": "Listar buckets",
        "operationId": "listBuckets",
        "responses": {
          "200": {
            "description": "Lista de buckets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Bucket" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Buckets"],
        "summary": "Criar bucket",
        "operationId": "createBucket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BucketCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bucket criado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bucket" }
              }
            }
          },
          "400": {
            "description": "Dados invalidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}": {
      "get": {
        "tags": ["Buckets"],
        "summary": "Obter bucket por ID",
        "operationId": "getBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Bucket encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bucket" }
              }
            }
          },
          "404": {
            "description": "Bucket nao encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Buckets"],
        "summary": "Atualizar bucket",
        "operationId": "updateBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BucketUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bucket atualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bucket" }
              }
            }
          },
          "400": {
            "description": "Dados invalidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Bucket nao encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Buckets"],
        "summary": "Excluir bucket",
        "operationId": "deleteBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Bucket excluido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bucket deve estar parado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Bucket nao encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/stats": {
      "get": {
        "tags": ["Buckets"],
        "summary": "Estatisticas do bucket",
        "operationId": "getBucketStats",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Estatisticas",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Stats" }
              }
            }
          },
          "404": {
            "description": "Bucket nao encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/folders": {
      "get": {
        "tags": ["Buckets"],
        "summary": "Estatisticas por pasta de origem",
        "operationId": "getBucketFolders",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Lista de pastas com estatisticas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "source_folder": { "type": "string" },
                      "pending": { "$ref": "#/components/schemas/StatusCount" },
                      "in_progress": { "$ref": "#/components/schemas/StatusCount" },
                      "completed": { "$ref": "#/components/schemas/StatusCount" },
                      "error": { "$ref": "#/components/schemas/StatusCount" },
                      "conflict": { "$ref": "#/components/schemas/StatusCount" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bucket nao encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/files/{status}": {
      "get": {
        "tags": ["Files"],
        "summary": "Listar arquivos do bucket por status",
        "operationId": "getBucketFiles",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" },
          { "$ref": "#/components/parameters/FileStatus" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": {
            "description": "Lista de arquivos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FileEntry" }
                }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/activity": {
      "get": {
        "tags": ["Buckets"],
        "summary": "Atividade recente do bucket",
        "operationId": "getBucketActivity",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 10000 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de atividades recentes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FileEntry" }
                }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/service": {
      "get": {
        "tags": ["Service Control"],
        "summary": "Status do servico do bucket",
        "operationId": "getBucketServiceStatus",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Status do servico",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceStatus" }
              }
            }
          },
          "404": {
            "description": "Bucket nao encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/service/start": {
      "post": {
        "tags": ["Service Control"],
        "summary": "Iniciar servico do bucket",
        "operationId": "startBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Servico iniciado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceStatus" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/service/pause": {
      "post": {
        "tags": ["Service Control"],
        "summary": "Pausar servico do bucket",
        "operationId": "pauseBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Servico pausado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceStatus" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/service/resume": {
      "post": {
        "tags": ["Service Control"],
        "summary": "Retomar servico do bucket",
        "operationId": "resumeBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Servico retomado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceStatus" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/service/stop": {
      "post": {
        "tags": ["Service Control"],
        "summary": "Parar servico do bucket",
        "operationId": "stopBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Servico parado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceStatus" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/service/workers": {
      "post": {
        "tags": ["Service Control"],
        "summary": "Alterar numero de workers",
        "operationId": "setWorkerCount",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WorkerCount" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workers atualizados",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceStatus" }
              }
            }
          },
          "400": {
            "description": "Count invalido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/scan": {
      "post": {
        "tags": ["Scanner"],
        "summary": "Varrer pastas de origem do bucket",
        "operationId": "scanBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Resultado da varredura",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanResult" }
              }
            }
          },
          "404": {
            "description": "Bucket nao encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/conflicts/{fileId}/resolve": {
      "post": {
        "tags": ["Files"],
        "summary": "Resolver conflito de arquivo no bucket",
        "operationId": "resolveConflictForBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" },
          { "$ref": "#/components/parameters/FileId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConflictResolution" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conflito resolvido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "changes": { "type": "integer" } }
                }
              }
            }
          },
          "400": {
            "description": "Action invalida",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/conflicts/resolve-all": {
      "post": {
        "tags": ["Files"],
        "summary": "Resolver todos os conflitos do bucket",
        "operationId": "resolveAllConflictsForBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConflictResolution" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conflitos resolvidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "changes": { "type": "integer" } }
                }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/errors/{fileId}/retry": {
      "post": {
        "tags": ["Files"],
        "summary": "Retentar arquivo com erro no bucket",
        "operationId": "retryErrorForBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" },
          { "$ref": "#/components/parameters/FileId" }
        ],
        "responses": {
          "200": {
            "description": "Arquivo retornado para fila",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "changes": { "type": "integer" } }
                }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/errors/retry-all": {
      "post": {
        "tags": ["Files"],
        "summary": "Retentar todos os erros do bucket",
        "operationId": "retryAllErrorsForBucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" }
        ],
        "responses": {
          "200": {
            "description": "Erros retornados para fila",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "changes": { "type": "integer" } }
                }
              }
            }
          }
        }
      }
    },
    "/buckets/{id}/export/{status}": {
      "get": {
        "tags": ["Files"],
        "summary": "Exportar arquivos do bucket em CSV",
        "operationId": "exportBucketFiles",
        "parameters": [
          { "$ref": "#/components/parameters/BucketId" },
          { "$ref": "#/components/parameters/FileStatus" }
        ],
        "responses": {
          "200": {
            "description": "Arquivo CSV",
            "content": {
              "text/csv": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
...
    "/files/{status}": {
      "get": {
        "tags": ["Files"],
        "summary": "Listar arquivos globais por status",
        "operationId": "getFilesByStatus",
        "parameters": [
          { "$ref": "#/components/parameters/FileStatus" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": {
            "description": "Lista de arquivos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FileEntry" }
                }
              }
            }
          }
        }
      }
    },
    "/conflicts/{fileId}/resolve": {
      "post": {
        "tags": ["Files"],
        "summary": "Resolver conflito de arquivo (global)",
        "operationId": "resolveConflict",
        "parameters": [
          { "$ref": "#/components/parameters/FileId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConflictResolution" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conflito resolvido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "changes": { "type": "integer" } }
                }
              }
            }
          }
        }
      }
    },
    "/conflicts/resolve-all": {
      "post": {
        "tags": ["Files"],
        "summary": "Resolver todos os conflitos (global)",
        "operationId": "resolveAllConflicts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConflictResolution" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conflitos resolvidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "changes": { "type": "integer" } }
                }
              }
            }
          }
        }
      }
    },
    "/errors/{fileId}/retry": {
      "post": {
        "tags": ["Files"],
        "summary": "Retentar arquivo com erro (global)",
        "operationId": "retryError",
        "parameters": [
          { "$ref": "#/components/parameters/FileId" }
        ],
        "responses": {
          "200": {
            "description": "Arquivo retornado para fila",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "changes": { "type": "integer" } }
                }
              }
            }
          }
        }
      }
    },
...
    "/export/{status}": {
      "get": {
        "tags": ["Files"],
        "summary": "Exportar todos os arquivos por status em CSV",
        "operationId": "exportFiles",
        "parameters": [
          { "$ref": "#/components/parameters/FileStatus" }
        ],
        "responses": {
          "200": {
            "description": "Arquivo CSV",
            "content": {
              "text/csv": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "BucketId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 },
        "description": "ID do bucket"
      },
...
    },
    "schemas": {
      "Bucket": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "source_folders": { "type": "array", "items": { "type": "string" } },
          "destination_folder": { "type": "string" },
          "worker_count": { "type": "integer" },
          "status": { "type": "string", "enum": ["stopped", "running", "paused"] },
          "created_at": { "type": "string" },
          "updated_at": { "type": "string" },
          "poolStatus": {
            "type": "object",
            "properties": {
              "bucketId": { "type": "integer" },
              "status": { "type": "string" },
              "workerCount": { "type": "integer" },
              "activeWorkers": { "type": "integer" }
            }
          }
        }
      },
      "BucketCreate": {
        "type": "object",
        "required": ["name", "destinationFolder"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "sourceFolders": { "type": "array", "items": { "type": "string" } },
          "destinationFolder": { "type": "string", "minLength": 1 },
          "workerCount": { "type": "integer", "minimum": 1, "maximum": 16 }
        }
      },
...
      "StatusCount": {
        "type": "object",
        "properties": {
          "count": { "type": "integer" },
          "totalSize": { "type": "integer" }
        }
      },
...
    }
  },
...
}